Grocery Stores in Census Tracts in Hamilton

This is a sf dataframe containing the information about the number of grocery stores in and census information about each Census Tract in Hamilton

Format

A sf dataframe with 34 columns:

GeoUID

The unique ID for the geographical unit

v_CA21_1

The population count in the census tract from 2021 Canadian Census

v_CA21_3

Population percentage change from 2016 to 2021

v_CA21_7

Land area in square kilometres

v_CA21_8

Total number of population in the census track with a valid age value

v_CA21_11

Total number of population between 0 to 14 years old

v_CA21_10

Total number of female population in the census track with a valid age value

v_CA21_251

Total number of population aged 65 years old or above

v_CA21_71

Total number of population aged 15 to 19 years old

v_CA21_89

Total number of population aged 20 to 24 years old

v_CA21_435

Total number of population live in single detached house

v_CA21_434

Total number of population with a valid value in occupied private dwelling by structural type of dwelling data

v_CA21_453

Total number of population with a valid value in marital status for the total population aged 15 years and over

v_CA21_456

Total number of population married or living common-law

v_CA21_1142

Gini index on adjusted household after tax income

v_CA21_1144

Total number of population with a valid value in knowledge of official languages excluding institutional residents

v_CA21_671

Total number of population with total income in 2020

v_CA21_674

Total number of population with income less than 10,000

v_CA21_677

Total number of population with income from 10,000 to 19,999

v_CA21_680

Total number of population with income from 20,000 to 29,999

v_CA21_683

Total number of population with income from 30,000 to 39,999

v_CA21_686

Total number of population with income from 40,000 to 49,999

v_CA21_689

Total number of population with income from 50,000 to 59,999

v_CA21_692

Total number of population with income from 60,000 to 69,999

v_CA21_695

Total number of population with income from 70,000 to 79,999

v_CA21_698

Total number of population with income from 80,000 to 89,999

v_CA21_701

Total number of population with income from 90,000 to 99,999

v_CA21_707

Total number of population with income from 100,000 to 149,999

v_CA21_710

Total number of population with income from 150,000 and over

v_CA21_1156

Total number of population with no knowledge of neither English or French

v_CA21_2176

Total number of population whose language spoken at home is non-official language

v_CA21_2167

Total number of population with a valid value in all languages spoken at home excluding institutional residents

Freq

The number of grocery stores in that census tract

geometry

Geometry column for sf dataframe

Examples

data(grocery_CT)
summary(grocery_CT[,c("GeoUID", "Freq")])
#>     GeoUID               Freq                 geometry  
#>  Length:150         Min.   :0.0000   MULTIPOLYGON :150  
#>  Class :character   1st Qu.:0.0000   epsg:4326    :  0  
#>  Mode  :character   Median :0.0000   +proj=long...:  0  
#>                     Mean   :0.5867                      
#>                     3rd Qu.:1.0000                      
#>                     Max.   :4.0000                      
head(grocery_CT[,c("GeoUID", "Freq")])
#> Simple feature collection with 6 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -79.89331 ymin: 43.17916 xmax: -79.80518 ymax: 43.21881
#> Geodetic CRS:  WGS 84
#>       GeoUID Freq                       geometry
#> 1 5370001.01    2 MULTIPOLYGON (((-79.82606 4...
#> 2 5370001.02    1 MULTIPOLYGON (((-79.85379 4...
#> 3 5370001.04    0 MULTIPOLYGON (((-79.86939 4...
#> 4 5370001.05    1 MULTIPOLYGON (((-79.8892 43...
#> 5 5370001.06    0 MULTIPOLYGON (((-79.8504 43...
#> 6 5370001.07    0 MULTIPOLYGON (((-79.83722 4...