Grocery Stores in Census Tracts in Hamilton

This is a sf dataframe containing the information about the number of grocery stores in each Census Tract in Hamilton

Format

A sf dataframe with 4 columns:

GEOUID

The unique ID for the geographical unit

POP21

The population count in the census tract from 2021 Canadian Census

Grocery

The number of grocery stores in that census tract

geometry

Geometry column for sf dataframe

Examples

data(grocery_CT)
summary(grocery_CT[,c("GEOUID", "POP21", "Grocery")])
#>     GEOUID              POP21         Grocery      
#>  Length:150         Min.   :  53   Min.   :0.0000  
#>  Class :character   1st Qu.:2849   1st Qu.:0.0000  
#>  Mode  :character   Median :3549   Median :0.0000  
#>                     Mean   :3796   Mean   :0.5867  
#>                     3rd Qu.:4672   3rd Qu.:1.0000  
#>                     Max.   :8587   Max.   :4.0000  
head(grocery_CT[,c("GEOUID", "POP21", "Grocery")])
#>       GEOUID POP21 Grocery
#> 1 5370001.01  2133       2
#> 2 5370001.02  5366       1
#> 3 5370001.04  6603       0
#> 4 5370001.05  5697       1
#> 5 5370001.06  5154       0
#> 6 5370001.07  3621       0